libs/mDNS: Reorganize
Use a clean tarball from the mdns github project to make the origin
more transparent and simplify updates.

Split out the android stuff in separate dir.
