+++ libs/mDNS/include/android/LocalArray.h
