+++ libs/mDNS/include/android/ScopedFd.h
