+++ libs/mDNS/include/android/ifaddrs-android.h
