+++ libs/mDNS/include/mdns_util.h
@@ -21,9 +21,11 @@
 extern mdns_string_t
 ipv4_address_to_string(char* buffer, size_t capacity, const struct sockaddr_in* addr,
                        size_t addrlen);
+
 extern mdns_string_t
 ipv6_address_to_string(char* buffer, size_t capacity, const struct sockaddr_in6* addr,
                        size_t addrlen);
+
 extern mdns_string_t
 ip_address_to_string(char* buffer, size_t capacity, const struct sockaddr* addr, size_t addrlen);
