+++ libs/mDNS/mdns-1.4.3/.github/FUNDING.yml
@@ -1,0 +1,3 @@
+# These are supported funding model platforms
+
+github: [mjansson]
